+++ src/main/java/com/qindesign/json/schema/Main.java
@@ -177,8 +177,8 @@ public class Main {
     root.add("valid", new JsonPrimitive(result));
     JsonArray errorArr = new JsonArray();
     root.add("errors", errorArr);
-    errors.forEach((schemaPath, map) -> {
-      map.forEach((instancePath, a) -> {
+    errors.forEach((instancePath, map) -> {
+      map.forEach((schemaPath, a) -> {
         JsonObject error = new JsonObject();
         error.add("keywordLocation", new JsonPrimitive(a.keywordLocation.toString()));
         error.add("absoluteKeywordLocation", new JsonPrimitive(a.absKeywordLocation.toString()));