+++ src/main/java/com/qindesign/json/schema/Validator.java
@@ -187,14 +187,14 @@ public final class Validator {
    *     any default</li>
    * </ol>
    * <p>
-   * The annotations and errors are maps from instance or schema locations to an
+   * The annotations and errors are maps from instance locations to an
    * associated {@link Annotation}, with some intervening values. Locations are
    * given as <a href="https://tools.ietf.org/html/rfc6901">JSON Pointers</a>.
    * <ul>
    * <li>The annotations follow this structure: instance location &rarr; name
    *     &rarr; schema location &rarr; {@link Annotation}. The
    *     {@link Annotation} value is dependent on the source.</li>
-   * <li>The errors have this structure: schema location &rarr; instance
+   * <li>The errors have this structure: instance location &rarr; schema
    *     location &rarr; {@link Annotation}. The {@link Annotation} value is an
    *     instance of {@link ValidationResult}, and its name will be "error" when
    *     the result is {@code false} and "annotation" when the result is